Adjust PlayPoint migration. Fix typo in Equip Migration :^)
