+++ Arrowgene.Ddon.Database/Arrowgene.Ddon.Database.csproj
@@ -50,5 +50,11 @@
         <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
       </None>
     </ItemGroup>
+	
+	<ItemGroup>
+      <None Update="Files\Database\Script\migration_pp_sqlite.sql">
+        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+      </None>
+    </ItemGroup>
 
 </Project>
