+++ Arrowgene.Ddon.Database/Files/Database/Script/schema_sqlite.sql
@@ -551,3 +551,12 @@ CREATE TABLE IF NOT EXISTS "ddon_system_mail_attachment" (
     PRIMARY KEY("attachment_id" AUTOINCREMENT),
 	FOREIGN KEY("message_id") REFERENCES "ddon_system_mail"("message_id") ON DELETE CASCADE
 );
+
+CREATE TABLE IF NOT EXISTS "ddon_character_playpoint_data" (
+	"character_common_id"	INTEGER NOT NULL,
+	"job"					SMALLINT NOT NULL,
+	"play_point"			INTEGER NOT NULL DEFAULT 0,
+	"exp_mode"				TINYINT NOT NULL DEFAULT 1,
+	CONSTRAINT pk_character_playpoint PRIMARY KEY (character_common_id, job),
+	CONSTRAINT fk_character_playpoint_character_common_id FOREIGN KEY("character_common_id") REFERENCES "ddon_character_common"("character_common_id") ON DELETE CASCADE
+);
