+++ Arrowgene.Ddon.Database/Sql/Core/Migration/00000004_PlayPointMigration.cs
@@ -1,0 +1,25 @@
+using System.Data.Common;
+
+namespace Arrowgene.Ddon.Database.Sql.Core.Migration
+{
+    public class PPMigration : IMigrationStrategy
+    {
+        public uint From => 3; 
+        public uint To => 4;
+
+        private readonly DatabaseSetting DatabaseSetting;
+
+        public PPMigration(DatabaseSetting databaseSetting)
+        {
+            DatabaseSetting = databaseSetting;
+        }
+
+        public bool Migrate(IDatabase db, DbConnection conn)
+        {
+            string adaptedSchema = DdonDatabaseBuilder.GetAdaptedSchema(DatabaseSetting, "Script/migration_pp_sqlite.sql");
+            db.Execute(conn, adaptedSchema);
+
+            return true;
+        }
+    }
+}
